+++ src/components/Card/Card.tsx
@@ -1,12 +1,12 @@
-import { PropsWithChildren, createContext, useContext, useRef, useState } from 'react';
+import { PropsWithChildren, createContext, useContext } from 'react';
 import Image, { StaticImageData } from 'next/image';
 import classNames from 'classnames/bind';
 
-import Button from '@/components/Button/Button';
 import styles from './Card.module.scss';
-import useClickOutside from '@/hooks/useClickOutside';
 import { RESERVATION_STATE_LABEL_MAP } from '@/constants';
-import { StarIcon, MeatballIcon } from '@/images/icon';
+import { StarIcon } from '@/images/icon';
+import { CardDropdown } from './Components/CardDropdown';
+import { ReservationButton } from './Components/ReservationButton';
 
 type ReservationState = 'completed' | 'canceled' | 'rejected' | 'finished';
 
@@ -37,9 +37,11 @@ export function ReservationCard({ card }: { card: CardProps }) {
     <Card card={card}>
       <Card.Thumbnail />
       <Card.Description>
-        <Card.ReservationState />
-        <Card.Title />
-        <Card.Schedule />
+        <div>
+          <Card.ReservationState />
+          <Card.Title />
+          <Card.Schedule />
+        </div>
         <Card.Footer>
           <Card.Price />
           <Card.ReservationButton />
@@ -55,8 +57,10 @@ export function ExperienceCard({ card }: { card: CardProps }) {
     <Card card={card}>
       <Card.Thumbnail />
       <Card.Description>
-        <Card.StarRating />
-        <Card.Title />
+        <div>
+          <Card.StarRating />
+          <Card.Title />
+        </div>
         <Card.Footer>
           <Card.Price />
           <Card.Dropdown />
@@ -84,7 +88,7 @@ export function ReviewCard({ card }: { card: CardProps }) {
 const cn = classNames.bind(styles);
 const CardContext = createContext<CardProps>({} as CardProps);
 
-function useCard() {
+export function useCard() {
   return useContext(CardContext);
 }
 
@@ -105,7 +109,7 @@ Card.Divider = Divider;
 Card.Footer = Footer;
 Card.Price = Price;
 Card.ReservationButton = ReservationButton;
-Card.Dropdown = Dropdown;
+Card.Dropdown = CardDropdown;
 
 function Thumbnail({ where }: { where?: 'review' }) {
   const { imgUrl } = useCard();
@@ -168,73 +172,4 @@ function Price({ where }: { where?: 'review' }) {
   const formattedPrice = new Intl.NumberFormat('ko-KR').format(Number(price));
 
   return <div className={cn('price', where)}>₩{formattedPrice}</div>;
-}
-
-function ReservationButton() {
-  const { reservationState } = useCard();
-
-  const handleReservationCancelClick = () => {
-    // 예약취소 구현
-  };
-
-  const handleReviewCreateClick = () => {
-    // 후기작성 구현
-  };
-
-  if (reservationState === 'completed') {
-    return (
-      <Button className={cn('reservationButton')} type="secondary" size="medium" onClick={handleReservationCancelClick}>
-        예약 취소
-      </Button>
-    );
-  }
-
-  if (reservationState === 'finished') {
-    return (
-      <Button className={cn('reservationButton')} type="primary" size="medium" onClick={handleReviewCreateClick}>
-        후기 작성
-      </Button>
-    );
-  }
-
-  return null;
-}
-
-function Dropdown() {
-  const [isOpenDropdown, setViewDropdown] = useState(false);
-  const profileRef = useRef<HTMLDivElement>(null);
-
-  const toggleDropdown = () => {
-    setViewDropdown(!isOpenDropdown);
-  };
-
-  const closeDropdown = () => {
-    setViewDropdown(false);
-  };
-
-  const handleModifyClick = () => {
-    // 수정하기 구현
-  };
-
-  const handleDeleteClick = () => {
-    // 삭제하기 구현
-  };
-
-  useClickOutside(profileRef, closeDropdown);
-
-  return (
-    <div className={cn('dropdownContainer')} ref={profileRef}>
-      <MeatballIcon className={cn('meatball')} viewBox="0 0 40 40" onClick={toggleDropdown} />
-      {isOpenDropdown && (
-        <div className={cn('dropdown')}>
-          <div className={cn('dropdownItem')} onClick={handleModifyClick}>
-            수정하기
-          </div>
-          <div className={cn('dropdownItem')} onClick={handleDeleteClick}>
-            삭제하기
-          </div>
-        </div>
-      )}
-    </div>
-  );
 }
